Start to write real fixtures in the tests/fixtures directory
