+++ tests/helpers/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+
+Project bw_processing developed at Funartech.
+Copyright © Funartech Inc. 2023
+"""
